debian/patches: Regenerate the patches from patch queue.
With this change, we are very close (completely ready?) to having mongodb
working on kFreeBSD (and, perhaps, even HURD in the not-so-distant future).
We *may* have to switch from v8 to spidermonkey, as a libv8 that is ported
to kFreeBSD is apparently only in experimental at the moment of this
writing.

Thanks: Pino Toscano and Jeff Epler
Closes: #703061