Generate dbgsym packages by default
Upstream's --nostrip option is not the same as Debian's nostrip
build option: our nostrip build option means that the
resulting binaries in the actual mongodb package should - as an
exception - be unstripped and is handled by dh_strip itself. Not passing
--nostrip to scons by default has a different effect: the binaries are
stripped during the build process, before dh_strip gets a chance to
generate -dbgsym packages.

We address this by changing the default to passing --nostrip to scons,
however because the -dbgsym packages are huge and their creation
time-consuming, allow a new "nodbgsym" flag in DEB_BUILD_OPTIONS to
negate this behavior.